CheckIn CheckOut Status Read and Update from Database using useEffect hook
This commit will replace the existing custom column to the one solution provided in github discussion. Also added two useEffect hooks, one to read the database when the row.id changes and the second to update the database when the CheckIn/Out button status changes. Issue: useEffect for the PATCH api call still tun on first render even using the useRef work around. Still need to fix that.